pyproject.toml
[tool.poetry]
name = "mobile-collab-robot"
version = "0.1.0"
description = ""
authors = ["wenliangyeoh <wlyeoh@cc.saga-u.ac.jp>"]
readme = "README.md"
packages = [{include = "mobile_collab_robot"}]

[tool.poetry.dependencies]
python = ">=3.11, <3.13"
numpy = "~2"
pykeigan-motor = "^2.4.0"
pyserial = "^3.5"
sounddevice = "^0.4.6"
psutil = "~5.9"
scipy = "^1.14.1"

[tool.poetry.group.dev.dependencies]
matplotlib = "~3.8"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
